// apps/admin/src/shared/api/generated/ja-korea/schemas/list6Params.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export type List6Params = {
published?: boolean;
title?: string;
authorName?: string;
publishFrom?: string;
publishTo?: string;
createdFrom?: string;
createdTo?: string;
page?: number;
size?: number;
};